fix: Redesign scoped context in jsonld (#750)
* Adds sub-@context to toplevel @context instead of to values

* Fixes @context missing @id property

* Annotates json-ld properties with type
